moved methods to score aggregation
