+++ sciencebeam_judge/evaluation_pipeline.py
@@ -44,11 +44,8 @@ from sciencebeam_gym.utils.file_list import (
 from sciencebeam_judge.evaluation_utils import (
   parse_xml,
   parse_xml_mapping,
-  score_results,
-  scoring_method_as_top_level_key,
-  combine_and_compact_scores_by_scoring_method_with_count,
-  summarise_results_by_scoring_method_with_count,
-  comma_separated_str_to_list
+  comma_separated_str_to_list,
+  score_results
 )
 
 from sciencebeam_judge.evaluation_config import (
@@ -63,6 +60,12 @@ from sciencebeam_judge.grobid_evaluate import (
 from .evaluation.scoring_methods import (
   ScoreMeasures,
   ALL_SCORE_MEASURES
+)
+
+from .evaluation.score_aggregation import (
+  scoring_method_as_top_level_key,
+  combine_and_compact_scores_by_scoring_method_with_count,
+  summarise_results_by_scoring_method_with_count
 )
 
 from .xpath_functions import register_functions
